[Go] track1 windowsesu rename package
