+++ services/preview/windowsesu/mgmt/2019-09-16-preview/windowsesu/operations.go
