+++ services/preview/windowsesu/mgmt/2019-09-16-preview/windowsesu/windowsesuapi/interfaces.go
@@ -8,7 +8,7 @@ package windowsesuapi
 
 import (
 	"context"
-	"github.com/Azure/azure-sdk-for-go/services/preview/windowsesu/2019-09-16-preview/windowsesu"
+	"github.com/Azure/azure-sdk-for-go/services/preview/windowsesu/mgmt/2019-09-16-preview/windowsesu"
 	"github.com/Azure/go-autorest/autorest"
 )
